Creates helper methods and test for tutorial sequencer
